+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Add icon, images, licenses, and descriptions for app store submission
 
+## [0.1.1] - 2021-08-26
+
+### Fixed
+
+- Fixed search query for notifications
+
+## [0.1.0] - 2021-08-25
+
+### Added
+
+- Add Locale and Seller
+
 ## [0.0.11] - 2021-07-14
 
 ### Added